pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "trafilatura>=2.0.0",
    "tweepy>=4.15.0",
]
